+++ config.json
@@ -1,0 +1,15 @@
+{
+    
+    "要升级的建筑物":[1,2,3],
+
+    "对应升级次数":[1,1,1],
+
+    "要收货的建筑物":[1,2,3],
+
+    "自动完成任务":"False",
+
+    "自动升级政策":"True",
+
+    "设备":"CB512BC4ZL"
+
+}
